move RSTReader
