+++ tests/rst/helpers.py
@@ -2,7 +2,7 @@
 from typing import List, Union
 
 # SOT
-from stubber.stubs_from_docs import RSTReader
+from stubber.rst.reader import RSTReader
 
 ###################################################################################################
 # Helpers
