+++ tests/rst/test_constants.py
@@ -5,7 +5,7 @@ import pytest
 from helpers import load_rst
 
 # SOT
-from stubber.stubs_from_docs import RSTReader
+from stubber.rst.reader import RSTReader
 
 # mark all tests
 pytestmark = pytest.mark.doc_stubs
